+++ zcash_client_sqlite/src/testing/db.rs
@@ -49,7 +49,6 @@ use {
     ::transparent::{address::TransparentAddress, bundle::OutPoint, keys::NonHardenedChildIndex},
     core::ops::Range,
     testing::transparent::GapLimits,
-    zcash_protocol::value::Zatoshis,
 };
 
 /// Tuesday, 25 February 2025 00:00:00Z (the day the clock code was added).